Export PIO output enables
OE signals can be used to determine if the ports are in high
impedance state.
Also remove the logic which automatically feeds back the output
register to input when the port is set to output, as it's possible
to 'override' the input even if the direction is output.

Feedback can be applied externally if desired, like:
IOx_in = IOx_oe ? IOx_out : port_input